+++ src/parser/block/paragraph.test.ts
@@ -29,7 +29,7 @@ describe('Unit: parser/block/paragraph', () => {
     it('mention', () => {
       assert.deepStrictEqual(inspect(parser('>a')), [['<p><a rel="noopener" class="address" data-level="1">&gt;a</a></p>'], '']);
       assert.deepStrictEqual(inspect(parser('>a\nb')), [['<p><a rel="noopener" class="address" data-level="1">&gt;a</a><br>b</p>'], '']);
-      assert.deepStrictEqual(inspect(parser('>a\nb\n>c')), [['<p><a rel="noopener" class="address" data-level="1">&gt;a</a><br>b<a rel="noopener" class="address" data-level="1">&gt;c</a></p>'], '']);
+      assert.deepStrictEqual(inspect(parser('>a\nb\n>c')), [['<p><a rel="noopener" class="address" data-level="1">&gt;a</a><br>b<br><a rel="noopener" class="address" data-level="1">&gt;c</a></p>'], '']);
       assert.deepStrictEqual(inspect(parser('>a\n>b')), [['<p><a rel="noopener" class="address" data-level="1">&gt;a</a><br><a rel="noopener" class="address" data-level="1">&gt;b</a></p>'], '']);
       assert.deepStrictEqual(inspect(parser('>a\n> b')), [['<p><a rel="noopener" class="address" data-level="1">&gt;a</a><br><span class="quotation">&gt; b</span></p>'], '']);
       assert.deepStrictEqual(inspect(parser('>a\n> b\nc')), [['<p><a rel="noopener" class="address" data-level="1">&gt;a</a><br><span class="quotation">&gt; b</span><br>c</p>'], '']);
@@ -39,8 +39,8 @@ describe('Unit: parser/block/paragraph', () => {
       assert.deepStrictEqual(inspect(parser('>a b\n>c')), [['<p><span class="quotation">&gt;a b<br>&gt;c</span></p>'], '']);
       assert.deepStrictEqual(inspect(parser('a>b')), [['<p>a&gt;b</p>'], '']);
       assert.deepStrictEqual(inspect(parser('a >b')), [['<p>a &gt;b</p>'], '']);
-      assert.deepStrictEqual(inspect(parser('a\n>b')), [['<p>a<a rel="noopener" class="address" data-level="1">&gt;b</a></p>'], '']);
-      assert.deepStrictEqual(inspect(parser('a\n>b\nc')), [['<p>a<a rel="noopener" class="address" data-level="1">&gt;b</a><br>c</p>'], '']);
+      assert.deepStrictEqual(inspect(parser('a\n>b')), [['<p>a<br><a rel="noopener" class="address" data-level="1">&gt;b</a></p>'], '']);
+      assert.deepStrictEqual(inspect(parser('a\n>b\nc')), [['<p>a<br><a rel="noopener" class="address" data-level="1">&gt;b</a><br>c</p>'], '']);
       assert.deepStrictEqual(inspect(parser(' >a')), [['<p>&gt;a</p>'], '']);
     });