+++ src/parser/block/paragraph.ts
@@ -4,13 +4,25 @@ import { mention } from './paragraph/mention';
 import { inline } from '../inline';
 import { contentline } from '../source';
 import { defrag, hasContent } from '../util';
+import { concat } from 'spica/concat';
 import { html } from 'typed-dom';
 
 export const paragraph: ParagraphParser = block(fmap(
   some(subsequence([
-    some(mention),
-    rewrite(
-      some(contentline, /^>/),
-      defrag(trim(some(inline)))),
+    fmap(
+      some(mention),
+      es => es.reduce((acc, el) => concat(acc, [el, html('br')]), [])),
+    fmap(
+      rewrite(
+        some(contentline, /^>/),
+        defrag(trim(some(inline)))),
+      ns => concat(ns, [html('br')])),
   ])),
-  ns => [html('p', ns)].filter(hasContent)));
+  ns => [html('p', format(ns))].filter(hasContent)));
+
+function format<T extends Node[]>(ns: T): T {
+  ns[ns.length - 1] instanceof HTMLBRElement && void ns.pop();
+  assert(!(ns[0] instanceof HTMLBRElement));
+  assert(!(ns[ns.length - 1] instanceof HTMLBRElement));
+  return ns;
+}